pyproject.toml
[project]
name = "kafka_collector"
version = "0.0.1"
description = "collect kafka messages from multiple topics"
authors = [
    {name = "Siak Hooi",email = "siakhooi@gmail.com"}
]
keywords = [ "kafka" ]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Topic :: Utilities"
]

license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kafka-python-ng>=2.2.0",
]

[project.urls]
homepage = "https://github.com/siakhooi/kafka-collector"
repository = "https://github.com/siakhooi/kafka-collector"
documentation = "https://github.com/siakhooi/kafka-collector/wiki"
"Bug Tracker" = "https://github.com/siakhooi/kafka-collector/issues"

[project.scripts]
kafka-collector = 'kafka_collector.cli:run'


[tool.poetry]
packages = [{include = "kafka_collector", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.1.0"
flake8 = "^7.2.0"
pytest-mock = "^3.15.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
